requirements.txt
requests
flask
python-dotenv
python-csv